05-23
用户加载已分配和未分配的案件列表
根据案件id查看案件关联人和案件详情页面

+++ deptone/src/main/java/com/deptcase/casemgt/entity/ContactRecordPo.java
@@ -5,7 +5,7 @@ public class ContactRecordPo {
 
     private Long ccId;
 
-    private Integer caseId;
+    private Long caseId;
 
     private String caseContact;
 
@@ -33,11 +33,11 @@ public class ContactRecordPo {
         this.ccId = ccId;
     }
 
-    public Integer getCaseId() {
+    public Long getCaseId() {
         return caseId;
     }
 
-    public void setCaseId(Integer caseId) {
+    public void setCaseId(Long caseId) {
         this.caseId = caseId;
     }
